Re-structure data structures
1. Fix cookie out-of-range bug
2. Move string_view to string_view.h
3. Rename Node to List
